renamed subreducer #804
