+++ visualization/app/codeCharta/state/store/reducer.ts
@@ -1,7 +1,7 @@
 import appSettings from "./appSettings/appSettings.reducer"
 import fileSettings from "./fileSettings/fileSettings.reducer"
 import dynamicSettings from "./dynamicSettings/dynamicSettings.reducer"
-import treeMap from "./treeMap/mapSize.reducer"
+import treeMap from "./treeMap/treeMap.reducer"
 import { combineReducers } from "redux"
 
 const rootReducer = combineReducers({
